Spacemacs
Now using Spacemacs as a base for emacs. Symlinking .spacemacs file for home
directory and other configuration files are loaded from .spacemacs. This means
much fewer files need to be tracked e.g. packages, and it's faster to set up.
Spacemacs so far has all of the features that was in the previous emacs configs
with a lot fewer bugs.